Strip unnecessary leading ha (#55)
In the HA CLI the prompt already indicates that we are running in the HA
CLI (similar to virsh or other commands with built-in shell). However,
in regular shell the ha command needs to be used, so that is what most
documentation says. Note that 'ha' is not necessary and strip it from
the command.