Update rlwrap to 0.46.1 to fix partially repeated input
First five commands of every command are repeated on enter, the length
changes on the length of the `-S` argument passed to `rlwrap`. This
seems to be related to changes for readline v0.8.1/v0.8.2, as discussed
in a similar issue: https://github.com/hanslub42/rlwrap/issues/168
With latest rlwrap release the issue is gone, fixes #127.

Also note that the latest release is missing `v` prefix in the git tag,
thus it's been removed from the URL of the release tarball.